view разнесен по классам

+++ src/main/java/ua/edu/sumdu/j2se/lytovka/tasks/view/intervalHHMM.java
@@ -1,7 +1,7 @@
 package ua.edu.sumdu.j2se.lytovka.tasks.view;
 
 public class intervalHHMM {
-    public String get_intervalHHMM(int time) {
+    public String get_intervalHHMM(long time) {
         return String.format("%02d:%02d:%02d", time / 3600, time / 60 % 60, time % 60);
     }
 }
